dsbd_project/data_retrieval/requirements.txt
prometheus_api_client
flask
mysql.connector
